# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateElectricalTelemetryRequest:
    properties:
      charge_alt_volt:
        type: number
      earth_curr:
        type: number
      frequency:
        type: number
      l1_curr:
        type: number
      l1_l2_volt:
        type: number
      l1_n_volt:
        type: number
      l1_va:
        type: number
      l1_var:
        type: number
      l2_curr:
        type: number
      l2_l3_volt:
        type: number
      l2_n_volt:
        type: number
      l2_va:
        type: number
      l2_var:
        type: number
      l3_curr:
        type: number
      l3_l1_volt:
        type: number
      l3_n_volt:
        type: number
      l3_va:
        type: number
      l3_var:
        type: number
      percent_fp:
        type: number
      percent_fv:
        type: number
      pf_avg:
        type: number
      pf_l1:
        type: number
      pf_l2:
        type: number
      pf_l3:
        type: number
      total_va:
        type: number
      total_var:
        type: number
    type: object
  handler.CreateEngineTelemetryRequest:
    properties:
      avg_fuel_rate:
        type: number
      batt_volt:
        type: number
      coolant_temperature:
        type: number
      desired_operating_speed:
        type: integer
      ecu_temperature:
        type: number
      fuel_level_bottom:
        type: number
      fuel_level_pressure_1:
        type: number
      fuel_level_pressure_2:
        type: number
      fuel_level_top:
        type: number
      fuel_rate:
        type: number
      intake_manifold_pressure:
        type: number
      intake_manifold_temperature:
        type: number
      keyswitch_batt_potential:
        type: number
      oil_filter_out_pressure:
        type: number
      oil_pressure:
        type: number
      rated_power:
        type: number
      rated_speed:
        type: integer
      run_time:
        type: integer
      speed:
        type: integer
      total_fuel:
        type: number
      trip_fuel:
        type: number
      turbo_pressure:
        type: number
    type: object
...
  response.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  service.ElectricalTelemetryOutput:
    properties:
      charge_alt_volt:
        type: number
      created_at:
        type: string
      device_id:
        type: string
      earth_curr:
        type: number
      frequency:
        type: number
      l1_curr:
        type: number
      l1_l2_volt:
        type: number
      l1_n_volt:
        type: number
      l1_va:
        type: number
      l1_var:
        type: number
      l2_curr:
        type: number
      l2_l3_volt:
        type: number
      l2_n_volt:
        type: number
      l2_va:
        type: number
      l2_var:
        type: number
      l3_curr:
        type: number
      l3_l1_volt:
        type: number
      l3_n_volt:
        type: number
      l3_va:
        type: number
      l3_var:
        type: number
      percent_fp:
        type: number
      percent_fv:
        type: number
      pf_avg:
        type: number
      pf_l1:
        type: number
      pf_l2:
        type: number
      pf_l3:
        type: number
      total_va:
        type: number
      total_var:
        type: number
    type: object
  service.EngineTelemetryOutput:
    properties:
      avg_fuel_rate:
        type: number
      batt_volt:
        type: number
      coolant_temperature:
        type: number
      created_at:
        type: string
      desired_operating_speed:
        type: integer
      device_id:
        type: string
      ecu_temperature:
        type: number
      fuel_level_bottom:
        type: number
      fuel_level_pressure_1:
        type: number
      fuel_level_pressure_2:
        type: number
      fuel_level_top:
        type: number
      fuel_rate:
        type: number
      intake_manifold_pressure:
        type: number
      intake_manifold_temperature:
        type: number
      keyswitch_batt_potential:
        type: number
      oil_filter_out_pressure:
        type: number
      oil_pressure:
        type: number
      rated_power:
        type: number
      rated_speed:
        type: integer
      run_time:
        type: integer
      speed:
        type: integer
      total_fuel:
        type: number
      trip_fuel:
        type: number
      turbo_pressure:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@genset-monitoring.io
    name: API Support
  description: Production-ready IoT Genset Monitoring backend API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Genset Monitoring API
  version: "1.0"
paths:
...
  /api/v1/devices/{deviceID}/electrical:
    post:
      consumes:
      - application/json
      description: Registers new electrical telemetry for a specific device
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      - description: Electrical telemetry payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateElectricalTelemetryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Telemetry created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.ElectricalTelemetryOutput'
              type: object
        "400":
          description: Invalid UUID or request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Electrical Telemetry
      tags:
      - telemetry
  /api/v1/devices/{deviceID}/electrical/latest:
    get:
      description: Returns the most recent electrical telemetry for a specific device
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Latest telemetry
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.ElectricalTelemetryOutput'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Telemetry not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Latest Electrical Telemetry
      tags:
      - telemetry
  /api/v1/devices/{deviceID}/engine:
    post:
      consumes:
      - application/json
      description: Registers new engine telemetry for a specific device
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      - description: Engine telemetry payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateEngineTelemetryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Telemetry created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.EngineTelemetryOutput'
              type: object
        "400":
          description: Invalid UUID or request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Engine Telemetry
      tags:
      - telemetry
  /api/v1/devices/{deviceID}/engine/latest:
    get:
      description: Returns the most recent engine telemetry for a specific device
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Latest telemetry
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.EngineTelemetryOutput'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Telemetry not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Latest Engine Telemetry
      tags:
      - telemetry
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer eyJhbGciO..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"